Split backend system routes
